backend/requirements.txt
fastapi
uvicorn
python-multipart    # To handle file uploads
pdfminer.six
spacy
google-genai
python-dotenv